requirements.txt
accelerate>=0.20.0
apache-beam
bitsandbytes>=0.41.0
click
cmake
Cython
datasets>=2.12.0
Deprecated
diffusers==0.2.4
distlib
einops>=0.6.1
evaluate
gdown
gluonts
h11
h5py
hdfs
hopfield-layers @ git+https://github.com/ml-jku/hopfield-layers@f56f929c95b77a070ae675ea4f56b6d54d36e730
huggingface-hub
lazy_loader
matplotlib>=3.7.0
matplotlib-inline
multiprocess
nltk
numpy>=1.24.0
optuna
packaging
pandas>=2.0.0
Pillow
protobuf>=3.20.0
psutil
PyYAML
regex
requests
rich
safetensors>=0.3.1
scikit-image
scikit-learn>=1.2.0
scipy>=1.10.0
seaborn
sentencepiece>=0.1.99
six
tensorboard
timm @ git+https://github.com/rwightman/pytorch-image-models.git@7f19a4cce7004eee11704956d89c94566424f5ee
tokenizers
torch>=2.0.0
torchvision==0.10.0
tqdm>=4.65.0
transformers==4.31.0
wandb>=0.15.0
peft>=0.4.0